Thu, May 18, 2023, 2:34 PM +03:00

+++ finance/app.py
@@ -64,14 +64,14 @@ def buy():
         stocks = stock["name"]
         price = stock["price"]
 
-        total = float(price * shares)
+        total = price * shares
 
         if shares < 0:
             return apology("Shares Error")
 
         current_user = session["user_id"]
         cash = db.execute("SELECT cash FROM users where id = ?", current_user)
-        user_cash = cash[0]["cash"]
+        user_cash = float(cash[0]["cash"])
 
         if user_cash < total:
             return apology("Insufficient cash")
